order messages by date
Messages can have dates that do not correlate with their sequenceId.
Previously messages were accessed by sequenceId, this countered
any sorting. Now the sourceId is used for identification.

As this is a rather big change, this is going to be tested for a while
before it is being released to Google Play