do not break the app with one broken account
So far this only works for a scenario when a a unified account
with several real accounts is loaded
and of the real accounts is for some
reason broken, e.g. the password has changed.
In this case the app still loads the other accounts
and shows and indicator (on Android / Material).

To do:
- also start when there is just a single broken account
- display notification for iOS devices in some way